pyproject.toml
[tool.poetry]
name = "radio"
version = "0.1.0"
description = ""
authors = ["Ayaan Zaidi <hi@obviy.us>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.13"
yt-dlp = "^2022.8.14"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
